+++ src/Dosepy/app_controller/app_controller.py
@@ -3,17 +3,20 @@
 from PySide6.QtWidgets import (
     QFileDialog,
     QHeaderView,
+    QMessageBox,
 )
 from pathlib import Path
 from abc import ABC, abstractmethod
 import os
 import numpy as np
 
-from .app_components.file_dialog import (
+from Dosepy.app_components.file_dialog import (
     open_files_dialog,
     save_lut_file_dialog,
     Error_Dialog,
 )
+from Dosepy.tools.equate_files import equate, merge
+from Dosepy.image import stack_images, load
 
 class BaseController(ABC):
     """Abstract class."""
@@ -44,6 +47,7 @@ class CalibrationController(BaseController):
             if self._model.are_valid_tif_files(new_files):
                 current_files = self._view.cal_widget.get_files_list()
                 list_files = current_files + new_files
+
                 if self._model.are_files_equal_shape(list_files):
 
                     # Display path to files
@@ -68,9 +72,52 @@ class CalibrationController(BaseController):
                     self._view.cal_widget.apply_button.setEnabled(True)
                 
                 else:
-                    msg = "The tiff files must have the same shape."
+                    
+                    msg = "Do you want to equalize the files?"
                     print(msg)
-                    Error_Dialog(msg).exec()
+                    dlg = QMessageBox()
+                    dlg.setWindowTitle("Equate images")
+                    dlg.setText(msg)
+                    dlg.setStandardButtons(QMessageBox.StandardButton.Yes | QMessageBox.StandardButton.No)
+                    dlg.setIcon(QMessageBox.Icon.Question)
+                    button = dlg.exec()
+
+                    if button == QMessageBox.Yes:
+                        # Display path to files
+                        self._view.cal_widget.set_files_list(list_files)
+                        file_path = list_files[0]
+                        
+                        self._model.calibration_img = load(
+                            file_path,
+                            for_calib=True,
+                        )
+
+                        equal_images = equate(list_files)
+                        merged_images = merge(list_files, equal_images)
+
+                        img = stack_images(merged_images, padding=6)  # ArrayImage, but we need CalibrationImage
+                        self._model.calibration_img.array = img.array
+
+                        self._view.cal_widget.plot_image(self._model.calibration_img)
+
+                        # Find how many film we have and show a table for user input dose values
+                        self._model.calibration_img.set_labeled_img()
+                        num = self._model.calibration_img.number_of_films
+                        print(f"Number of detected films: {num}")
+                        self._view.cal_widget.set_table_rows(rows = num)
+                        header = self._view.cal_widget.dose_table.horizontalHeader()
+                        self._view.cal_widget.dose_table.cellChanged.connect(self._is_a_valid_dose)
+                        header.setSectionResizeMode(0, QHeaderView.ResizeMode.Stretch)
+
+                        self._view.cal_widget.apply_button.setEnabled(True)
+                        
+
+                    else:
+
+                        msg = "The tiff files must have the same shape."
+                        print(msg)
+                        Error_Dialog(msg).exec()
+
 
             else:
                 msg = "Invalid file. Is it a tiff RGB file?"
@@ -110,7 +157,7 @@ class CalibrationController(BaseController):
 
 
     def _save_calib_button(self):
-        root_calibration_path = Path(__file__).parent / "user" / "calibration"
+        root_calibration_path = Path(__file__).parent.parent / "user" / "calibration"
         if not root_calibration_path.exists():
             os.makedirs(root_calibration_path)
 
@@ -122,7 +169,7 @@ class CalibrationController(BaseController):
             print(lut_file_name)
         
             self._model.save_lut(str(lut_file_name))
-            #self._view.tif_widget.cali_label.setText(
+            #self._view.dose_widget.cali_label.setText(
             #    f"Calibration file: " + str(lut_file_name)
             #    )
 
@@ -166,18 +213,25 @@ class Tiff2DoseController(BaseController):
 
     
     def _open_tif2dose_button(self):
+        """
+        Uses a QFileDialog window to ask for tif files.
+        If files are ok, uses a calibration (ask if not exists) and calculate the dose distribution.
+        Show the dose distribution.
+        """
         
         new_files = open_files_dialog("Images (*.tif *.tiff)")
 
         if new_files:
     
             if self._model.are_valid_tif_files(new_files):
-                current_files = self._view.tif_widget.get_files_list()
+
+                current_files = self._view.dose_widget.get_files_list()
                 list_files = current_files + new_files
+                
                 if self._model.are_files_equal_shape(list_files):
 
                     # Display path to files
-                    self._view.tif_widget.set_files_list(list_files)
+                    self._view.dose_widget.set_files_list(list_files)
 
                     # load the files
                     self._model.tif_img = self._model.load_files(
@@ -185,7 +239,7 @@ class Tiff2DoseController(BaseController):
                         )
                     
                     if self._model.lut:
-                        self._model.dose = self._model.tif_img.to_dose(self._model.lut) # An ArrayImage
+                        self._model.ref_dose_img = self._model.tif_img.to_dose(self._model.lut) # An ArrayImage
 
                     else:
                         "Open lut"
@@ -197,7 +251,7 @@ class Tiff2DoseController(BaseController):
                         if lut_file_path:
                             if len(lut_file_path) == 1:
                                 self._model.lut = self._model.load_lut(lut_file_path[0])
-                                self._model.dose = self._model.tif_img.to_dose(
+                                self._model.ref_dose_img = self._model.tif_img.to_dose(
                                     self._model.lut
                                     ) # An ArrayImage
                             else:
@@ -205,14 +259,67 @@ class Tiff2DoseController(BaseController):
                                 print(msg)
                                 Error_Dialog(msg).exec()
                         else:
-                            self._view.tif_widget.files_list.clear()
+                            self._view.dose_widget.files_list.clear()
 
-                    self._view.tif_widget.plot_dose(self._model.dose)
+                    self._view.dose_widget.plot_dose(self._model.ref_dose_img)
 
                 else:
-                    msg = "The tiff files must have the same shape."
+                    
+                    msg = "Do you want to equalize the files?"
                     print(msg)
-                    Error_Dialog(msg).exec()
+                    dlg = QMessageBox()
+                    dlg.setWindowTitle("Equate images")
+                    dlg.setText(msg)
+                    dlg.setStandardButtons(QMessageBox.StandardButton.Yes | QMessageBox.StandardButton.No)
+                    dlg.setIcon(QMessageBox.Icon.Question)
+                    button = dlg.exec()
+
+                    if button == QMessageBox.Yes:
+                        # Display path to files
+                        #self._view.cal_widget.set_files_list(list_files)
+                        self._view.dose_widget.set_files_list(list_files)
+                        file_path = list_files[0]
+                        self._model.tif_img = load(file_path)  # Placeholder
+                        
+                        equal_images = equate(list_files, axis="width")
+
+                        merged_images = merge(list_files, equal_images)
+
+                        img = stack_images(merged_images, padding=6)
+                        # load the files
+                        self._model.tif_img.array = img.array 
+
+                        if self._model.lut:
+                            self._model.ref_dose_img = self._model.tif_img.to_dose(self._model.lut) # An ArrayImage
+
+                        else:
+                            "Open lut"
+                            lut_file_path = open_files_dialog(
+                                filter = "Calibration. (*.cal)",
+                                dir = "calib"
+                                )
+                            
+                            if lut_file_path:
+                                if len(lut_file_path) == 1:
+                                    self._model.lut = self._model.load_lut(lut_file_path[0])
+                                    self._model.ref_dose_img = self._model.tif_img.to_dose(
+                                        self._model.lut
+                                        ) # An ArrayImage
+                                else:
+                                    msg = "Chose one calibration file."
+                                    print(msg)
+                                    Error_Dialog(msg).exec()
+                            else:
+                                self._view.dose_widget.files_list.clear()
+
+                        self._view.dose_widget.plot_dose(self._model.ref_dose_img)
+                        
+
+                    else:
+
+                        msg = "The tiff files must have the same shape."
+                        print(msg)
+                        Error_Dialog(msg).exec()
             
             else:
                 msg = "Invalid file. Is it a tiff RGB file?"
@@ -242,6 +349,6 @@ class Tiff2DoseController(BaseController):
     
     def _connectSignalsAndSlots(self):
 
-        self._view.tif_widget.open_button.clicked.connect(self._open_tif2dose_button)
-        self._view.tif_widget.save_button.clicked.connect(self._save_tif2dose_button)
+        self._view.dose_widget.open_button.clicked.connect(self._open_tif2dose_button)
+        self._view.dose_widget.save_button.clicked.connect(self._save_tif2dose_button)